Refactoring common JSON/JsonPath functionality into JsonUtils
